refactor: makeLottos, makeExpectedWinningResult 함수를 testUtils/index.js로 옮김
